+++ packages/build-app-helpers/tests/fixtures/analyzeAuth/app.jsx
@@ -1,0 +1,9 @@
+import { runApp } from 'ice';
+
+runApp({
+  getInitialData: async () => {
+    return {
+      auth: {},
+    };
+  },
+});