+++ packages/build-app-helpers/tests/fixtures/analyzeAuth/app.tsx
@@ -1,0 +1,12 @@
+import { type AppConfig, runApp } from 'ice';
+
+const appConfig: AppConfig = {
+  getInitialData: function() {
+    return {};
+  },
+  auth: {
+    NoAuthFallback: <>no auth</>,
+  },
+}
+
+runApp(appConfig);