+++ packages/build-user-config/src/userConfig/mock.js
@@ -5,30 +5,33 @@ module.exports = (config, mock, context) => {
   const { commandArgs, command, webpack } = context;
   const isWebpack4 = /^4\./.test(webpack.version);
   if (!commandArgs.disableMock && command === 'start' && mock) {
+    // __FRAMEWORK_VERSION__ only defined in ice.js
+    // FIXME: remove process.env.__FRAMEWORK_VERSION__ when rax-app update @builder/pack
+    const hookName = process.env.__FRAMEWORK_VERSION__ ? 'setupMiddlewares' : 'onBeforeSetupMiddleware';
     // Compat with webpack4
-    const beforeHookName = isWebpack4 ? 'before' : 'onBeforeSetupMiddleware';
+    const beforeHookName = isWebpack4 ? 'before' : hookName;
     const originalDevServeBefore = config.devServer.get(beforeHookName);
     // replace devServer before function
     config.merge({ devServer: {
-      [beforeHookName](...args) {
+      [beforeHookName](appOrMiddlewares, server) {
+        let middlewares;
         let app;
-        let server;
-        if (args[0].app) {
-          server = args[0];
-          app = server.app;
+        if (appOrMiddlewares.get) {
+          app = appOrMiddlewares;
+        } else if (appOrMiddlewares.app) {
+          app = appOrMiddlewares.app;
         } else {
-          [app, server] = args;
+          app = server.app;
+          middlewares = appOrMiddlewares;
         }
-        // set cors before all served files
-        app.use((req, res, next) => {
-          res.set('Access-Control-Allow-Origin', '*');
-          next();
-        });
         const mockIgnore = Object.prototype.toString.call(mock) === '[object Object]' && mock.exclude;
         // keep mock server ahead of devServer.before
         webpackDevMock(app, mockIgnore || []);
         if (typeof originalDevServeBefore === 'function') {
-          originalDevServeBefore(...args);
+          return originalDevServeBefore(appOrMiddlewares, server);
+        }
+        if (middlewares) {
+          return middlewares;
         }
       },
     }});