+++ packages/create-app-shared/src/web/history.ts
@@ -3,17 +3,30 @@ import type { CreateHistory, InitHistory } from '../createInitHistory';
 import createInitHistory from '../createInitHistory';
 import { setHistory } from '../storage';
 
-const createHistory: CreateHistory = ({ type, basename, location }) => {
+const createHistory: CreateHistory = ({
+  type,
+  basename,
+  location,
+  customHistory,
+  initialIndex,
+  initialEntries,
+}) => {
   let history: History;
   if (process.env.__IS_SERVER__) {
     history = createMemoryHistory();
     (history as any).location = location;
-  } else if (type === 'hash') {
+  } else if (customHistory) {
+    history = customHistory;
+  }
+  else if (type === 'hash') {
     history = createHashHistory({ basename });
   } else if (type === 'browser') {
     history = createBrowserHistory({ basename });
   } else {
-    history = createMemoryHistory();
+    history = createMemoryHistory({
+      initialIndex,
+      initialEntries
+    });
   }
   setHistory(history);
   return history;