+++ packages/plugin-router/src/types/index.ts
@@ -12,6 +12,8 @@ export interface IAppRouterProps {
   modifyRoutes?: IModifyRoutes;
   fallback?: React.ReactNode;
   history?: History;
+  initialIndex?: number;
+  initialEntries?: string[];
 }
 
 export interface IRouterConfig extends DefaultRouteProps {