+++ packages/build-user-config/CHANGELOG.md
@@ -1,8 +1,23 @@
 # Changelog
 
-## 2.1.1
+## 2.2.3
 
 - [fix] revert postcss options merge
+
+## 2.2.2
+
+- [fix] compatible with rax-app devServer when config `webpack5`
+
+## 2.2.1
+
+- [fix] compatible with dev server
+
+## 2.2.0
+
+- [feat] support env variables
+- [feat] compatible with the latest API of `webpack-dev-server`
+- [fix] bump version of `@builder/pack`(^0.6.0)
+- [fix] remove default fallback for `url`
 
 ## 2.1.0