+++ packages/plugin-router/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.1.3
+
+- [fix] be able to pass arguments to memory router
+
+## 2.1.2
+
+- [fix] bump version of `@builder/pack`(^0.6.0)
+
 ## 2.1.1
 
 - [fix] bump version of `@builder/pack`(^0.5.0)