+++ packages/vite-service/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 2.1.1
+
+- [feat] plugin for rename main css chunk name
+- [fix] arguments of plugin-history may be `undefined` in case of `plugin-router` being deactivated
+- [fix] use `--disable-mock` command to escape MockPlugin
+
 ## 2.1.0
 
 - [feat] refactor html plugin for SSR render when development